Abstract signing multisig operations
