+++ packages/colony-starter-basic/src/helpers/executeTaskMultisig.js
@@ -1,0 +1,66 @@
+// A helper method for executing task multisig operations
+const executeTaskMultisig = async (colonyClient, taskMethod, taskId) => {
+
+  // Get JSON formatted operation from the mock database
+  const operationJSON = DATABASE.operations[taskMethod.name];
+
+  // Check the operation
+  if (operationJSON) {
+
+    // Restore operation
+    const operation = await taskMethod.restoreOperation(
+      operationJSON,
+    );
+
+    // Check if the operation matches the colony contract and task id and if
+    // the current account is one of the required signees for the operation.
+    if (
+      operation.payload.sourceAddress === colonyClient.contract.address &&
+      operation.payload.inputValues.taskId === taskId &&
+      operation.missingSignees.includes(colonyClient.adapter.wallet.address.toLowerCase())
+    ) {
+
+      // Sign the operation
+      await operation.sign();
+
+      // Check for missing signees
+      if (operation.missingSignees.length === 0) {
+
+        // Send the operation
+        await operation.send();
+
+        // Update the operation in the mock database
+        DATABASE.operations[taskMethod.name] = null;
+
+      } else {
+
+        // Check out the logs to see the operation missing signees
+        console.log('Missing Signees:', operation.missingSignees);
+
+        // Serialize operation into JSON format
+        const operationJSON = operation.toJSON();
+
+        // Store the operation in the mock database
+        DATABASE.operations[taskMethod.name] = operationJSON;
+
+      }
+
+    } else {
+
+      // Check out the logs to see the source of the pending operation
+      console.log('Operation Source:', operation.payload.sourceAddress);
+
+      // Check out the logs to see the task id of the pending operation
+      console.log('Operation Task ID:', operation.payload.inputValues.taskId);
+
+      // Check out the logs to see the required signees of the operation
+      console.log('Missing Signees:', operation.missingSignees);
+
+    }
+
+  }
+
+}
+
+// Export executeTaskMultisig example
+module.exports = executeTaskMultisig;
